add grouping fixes
