+++ src/components/revo-grid/revo-grid.tsx
@@ -17,7 +17,8 @@ import FilterPlugin, { ColumnFilterConfig, FilterCollection } from '../../plugin
 import SortingPlugin from '../../plugins/sorting/sorting.plugin';
 import ExportFilePlugin from '../../plugins/export/export.plugin';
 import { DataInput } from '../../plugins/export/types';
-import GroupingRowPlugin, { GroupingOptions } from '../../plugins/groupingRow/grouping.row.plugin';
+import GroupingRowPlugin from '../../plugins/groupingRow/grouping.row.plugin';
+import { GroupingOptions } from '../../plugins/groupingRow/grouping.row.types';
 
 type ColumnStores = {
   [T in RevoGrid.DimensionCols]: ObservableMap<DataSourceState<RevoGrid.ColumnRegular, RevoGrid.DimensionCols>>;
